Split up ast.go into more manageable pieces.
