+++ sql/parser2/delete.go
@@ -17,21 +17,15 @@
 
 package parser2
 
-import "bytes"
+import "fmt"
 
-// Truncate represents a TRUNCATE statement.
-type Truncate struct {
-	Tables []QualifiedName
+// Delete represents a DELETE statement.
+type Delete struct {
+	Table TableExpr
+	Where *Where
 }
 
-func (node *Truncate) String() string {
-	var buf bytes.Buffer
-	_, _ = buf.WriteString("TRUNCATE TABLE ")
-	for i, n := range node.Tables {
-		if i > 0 {
-			_, _ = buf.WriteString(", ")
-		}
-		_, _ = buf.WriteString(n.String())
-	}
-	return buf.String()
+func (node *Delete) String() string {
+	return fmt.Sprintf("DELETE FROM %v%v",
+		node.Table, node.Where)
 }
