+++ sql/parser2/drop.go
@@ -1,4 +1,4 @@
-// Copyright 2014 The Cockroach Authors.
+// Copyright 2015 The Cockroach Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,15 +17,7 @@
 
 package parser2
 
-import (
-	"bytes"
-	"fmt"
-)
-
-func (*DropDatabase) statement() {}
-func (*DropIndex) statement()    {}
-func (*DropTable) statement()    {}
-func (*DropView) statement()     {}
+import "bytes"
 
 // DropDatabase represents a DROP DATABASE statement.
 type DropDatabase struct {
@@ -35,52 +27,31 @@ type DropDatabase struct {
 
 func (node *DropDatabase) String() string {
 	var buf bytes.Buffer
-	buf.WriteString("DROP DATABASE ")
+	_, _ = buf.WriteString("DROP DATABASE ")
 	if node.IfExists {
-		buf.WriteString("IF EXISTS ")
+		_, _ = buf.WriteString("IF EXISTS ")
 	}
-	buf.WriteString(node.Name)
+	_, _ = buf.WriteString(node.Name)
 	return buf.String()
-}
-
-// DropIndex represents a DROP INDEX statement.
-type DropIndex struct {
-	Name      string
-	TableName string
-}
-
-func (node *DropIndex) String() string {
-	return fmt.Sprintf("DROP INDEX %s ON %s", node.Name, node.TableName)
 }
 
 // DropTable represents a DROP TABLE statement.
 type DropTable struct {
-	Name     string
+	Names    []QualifiedName
 	IfExists bool
 }
 
 func (node *DropTable) String() string {
 	var buf bytes.Buffer
-	buf.WriteString("DROP TABLE ")
+	_, _ = buf.WriteString("DROP TABLE ")
 	if node.IfExists {
-		buf.WriteString("IF EXISTS ")
+		_, _ = buf.WriteString("IF EXISTS ")
 	}
-	buf.WriteString(node.Name)
-	return buf.String()
-}
-
-// DropView represents a DROP TABLE statement.
-type DropView struct {
-	Name     string
-	IfExists bool
-}
-
-func (node *DropView) String() string {
-	var buf bytes.Buffer
-	buf.WriteString("DROP VIEW ")
-	if node.IfExists {
-		buf.WriteString("IF EXISTS ")
+	for i, n := range node.Names {
+		if i > 0 {
+			_, _ = buf.WriteString(", ")
+		}
+		_, _ = buf.WriteString(n.String())
 	}
-	buf.WriteString(node.Name)
 	return buf.String()
 }
